# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BookBridge",
    "description": "Read-only APIs over the public Books to Scrape HTML catalogue. Prices and ratings are fictional. Search scans a bounded number of pages and explicitly reports completeness. No credentials are required.",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "Operations"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/categories": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Get Categories",
        "operationId": "get_categories_v1_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryList"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/books": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Get Books",
        "description": "One upstream page (up to 20 books). Category IDs come from /v1/categories.",
        "operationId": "get_books_v1_books_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*_[1-9][0-9]*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookPage"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/v1/books/{book_id}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Get Book",
        "description": "Read public details, UPC, stock count, category and tax-inclusive price.",
        "operationId": "get_book_v1_books__book_id__get",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 250,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*_[1-9][0-9]*$",
              "title": "Book Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/v1/search": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "summary": "Search",
        "description": "Case-insensitive title substring search; at most five source pages per call.",
        "operationId": "search_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "start_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 1,
              "title": "Start Page"
            }
          },
          {
            "name": "max_pages",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "minimum": 1,
              "default": 3,
              "title": "Max Pages"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*_[1-9][0-9]*$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "upc": {
            "type": "string",
            "title": "Upc"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "stock_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Stock Count"
          },
          "price_excl_tax": {
            "$ref": "#/components/schemas/Price"
          },
          "tax": {
            "$ref": "#/components/schemas/Price"
          },
          "review_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Review Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "price",
          "rating",
          "in_stock",
          "source_url",
          "upc",
          "category",
          "description",
          "stock_count",
          "price_excl_tax",
          "tax",
          "review_count"
        ],
        "title": "Book"
      },
      "BookPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Total Pages"
          },
          "total_items": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Items"
          },
          "next_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page"
          },
          "previous_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Page"
          },
          "source": {
            "$ref": "#/components/schemas/SourceMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page",
          "total_pages",
          "total_items",
          "next_page",
          "previous_page",
          "source"
        ],
        "title": "BookPage"
      },
      "BookResult": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/Book"
          },
          "source": {
            "$ref": "#/components/schemas/SourceMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item",
          "source"
        ],
        "title": "BookResult"
      },
      "BookSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "price": {
            "$ref": "#/components/schemas/Price"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "in_stock": {
            "type": "boolean",
            "title": "In Stock"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "price",
          "rating",
          "in_stock",
          "source_url"
        ],
        "title": "BookSummary"
      },
      "Category": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "source_url"
        ],
        "title": "Category"
      },
      "CategoryList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Category"
            },
            "type": "array",
            "title": "Items"
          },
          "source": {
            "$ref": "#/components/schemas/SourceMeta"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "source"
        ],
        "title": "CategoryList"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "Price": {
        "properties": {
          "amount": {
            "type": "string",
            "pattern": "^\\d+\\.\\d{2}$",
            "title": "Amount",
            "examples": [
              "51.77"
            ]
          },
          "currency": {
            "type": "string",
            "const": "GBP",
            "title": "Currency",
            "default": "GBP"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount"
        ],
        "title": "Price"
      },
      "SearchResult": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "matched_in_scan": {
            "type": "integer",
            "title": "Matched In Scan"
          },
          "start_page": {
            "type": "integer",
            "title": "Start Page"
          },
          "scanned_pages": {
            "type": "integer",
            "title": "Scanned Pages"
          },
          "total_source_pages": {
            "type": "integer",
            "title": "Total Source Pages"
          },
          "next_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Page"
          },
          "complete": {
            "type": "boolean",
            "title": "Complete"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceMeta"
            },
            "type": "array",
            "title": "Sources"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": "Title substring matches in the scanned pages only; source order is preserved."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query",
          "items",
          "matched_in_scan",
          "start_page",
          "scanned_pages",
          "total_source_pages",
          "next_page",
          "complete",
          "sources"
        ],
        "title": "SearchResult"
      },
      "SourceMeta": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "live",
              "fixture"
            ],
            "title": "Mode"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mode",
          "source_url",
          "fetched_at",
          "cache_hit"
        ],
        "title": "SourceMeta"
      }
    }
  }
}